Add files via upload
The US housing market has changed as a result of the COVID-19 pandemic, affecting the cost and availability of single family housing. This project analyzes Zillow data to quantify 4 housing metrics: cost by state, cost by the top metro area in each state, availability, and time it takes for houses to go off-market.